feat: ✨ login modal
